feat: update redistribution address and abi (#3385)

+++ pkg/storageincentives/staking/abi.go
@@ -13,9 +13,9 @@ const ABIv0_0_0 = `[
         "type": "address"
       },
       {
-        "internalType": "uint8",
+        "internalType": "uint64",
         "name": "_NetworkId",
-        "type": "uint8"
+        "type": "uint64"
       }
     ],
     "stateMutability": "nonpayable",
